docs: Nightly validation fixes (#7249)
Looks like a docker link moved URLs.

A groovy example that failed snapshots after I properly committed the
iceberg config files. This is because 2 examples were trying to create
the same table in the same instance. Reset flag makes them get fresh
instances.

I am hoping the Python fix works. It's saying the example file doesn't
exist which is weird, but the snippet doesn't belong in a test set. It's
possible the iceberg config it was running with isn't including the
example data container. That's my current guess without spending too
much time on this.