fix: DH-20307: Throw errors when worker connection has been Disconnected (#7175)
DH-20307: Don't attempt to reconnect when connections are explicitly
closed.

This PR is necessary to support fixes made in
[deephaven-ent/iris/#3518](https://github.com/deephaven-ent/iris/pull/3518).
Specifically, there is a race condition where a created Core+ JS API
client can fail login and will continue to spam the server with login
attempts that will never succeed. That PR handles the client side piece
of this by explicitly disconnecting the client when this happens. The
server side fix is to stop attempting to reconnect if a connection is
explicitly closed.

### Testing
I have tested the 2 PRs together and have seen that this PR + the other
PR stops the rogue connection spamming. I have only seen change in
`connectToWorker` method of `WorkerConnection.java` actually get hit,
but I have also not observed any problems caused by it. (#DH-20307)