refactor: DH-20816: Refactor WhereFilter.Visitor pattern (#7382)
This change is removing the method
`WhereFilter.Visitor#visitWhereFilter(WhereFilter)`, which had a default
implementation using a bunch of instanceof checks which is antithetical
to the Visitor pattern, and was being "misused" by callers. While this
exact method signature could be reasonable, the current usage of it was
assuming this anti-pattern, and so a more explicit method name
`visitOther` was chosen to re-enforce the correct semantics - that is,
`visitOther` will only be called for filters that don't match a more
specifically typed method.

This refactoring has downstream effects, typically resulting in the call
to, or the implementation of, the Visitor relying on a Visitor-specific
static method to invoke the Visitor against a given WhereFilter.

In addition, some of the Visitor implementations were refactored to
improve their ergonomics or efficiency (ie, presenting Stream-based
implementation so aggregation can happen at the very end), migrating
Visitors to enum Singletons, removing unused code, and having callers
use more efficient Visitor-specific methods.

And finally, the method names `WhereFilter#walkWhereFilter` were changed
to `WhereFilter#walk` and the `WhereFilter.Visitor#visitWhereFilter`
were changed to `WhereFilter.Visitor#visit` to be more consistent with
the other visitor patterns in use.