fix: DH-20994: Corrected sort(float), sort(double) to fix chunk sorting issue.  (#7449)
https://deephaven.atlassian.net/browse/DH-20783 introducing new
functions to sort chunks based on Deephaven conventions. There was a bug
in sorting float/double where if the chunk only held NEGATIVE_INFINITY,
the next value outside of the specified range could be included in the
sort and would corrupt the results.

This PR addresses the issues and adds test to validate the correction is
in place.