fix: DH-21266: migrate lz4 dependency
The resolution block is necessary to work around a gradle bug, which otherwise causes the transitive lz4-java provided by kafka to conflict with the relocated version. This sort of workaround may not be necessary in the future once kafka has been updated to point to the new dependency, https://issues.apache.org/jira/browse/KAFKA-19951.

Fixes CVE-2025-12183 and CVE-2025-66566